+++ srcs/http/method/delete_method.cpp
@@ -8,7 +8,7 @@
 #include "http/const/const_html_filename.hpp"
 #include "http/const/const_response_key_map.hpp"
 #include "http/const/const_status_phrase.hpp"
-#include "http/http_parser_utils.hpp"
+#include "http/method/method_utils.hpp"
 #include "http/request/RequestInfo.hpp"
 #include "utils/file_io_utils.hpp"
